# docs/swagger.yaml
basePath: /
definitions:
...
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      role:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: suporte@seusite.com
    name: Suporte da API
    url: http://www.seusite.com/support
  description: API para gerenciamento de produtos, pedidos e usuários.
  title: API de Vendas
  version: "1.0"
paths:
...
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Permite que um usuário exclua sua conta
      parameters:
      - description: ID do usuário
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Usuário deletado com sucesso
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro ao deletar usuário
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Deleta um usuário
      tags:
      - Usuários
    get:
      consumes:
      - application/json
      description: Retorna informações do usuário pelo ID
      parameters:
      - description: ID do usuário
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "404":
          description: Usuário não encontrado
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Obtém os detalhes de um usuário
      tags:
      - Usuários
    put:
      consumes:
      - application/json
      description: Permite que o usuário atualize seu perfil com novas informações
      parameters:
      - description: ID do usuário
        in: path
        name: id
        required: true
        type: string
      - description: Dados do usuário para atualização
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: Perfil atualizado com sucesso
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Dados inválidos
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro ao atualizar perfil
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Atualiza os detalhes de um usuário
      tags:
      - Usuários
  /users/{id}/password:
    put:
      consumes:
      - application/json
      description: Permite que um usuário altere sua senha fornecendo a senha antiga
        e a nova senha
      parameters:
      - description: ID do usuário
        in: path
        name: id
        required: true
        type: string
      - description: Dados para alteração de senha
        in: body
        name: passwordData
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Senha alterada com sucesso
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Dados inválidos
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro ao alterar senha
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Altera a senha de um usuário
      tags:
      - Usuários
...
swagger: "2.0"